include attester as a file in the validator server
