+++ beacon-chain/rpc/validator/BUILD.bazel
@@ -3,6 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "attester.go",
+        "proposer.go",
         "assignments.go",
         "server.go",
         "status.go",
@@ -37,6 +39,7 @@ go_test(
     name = "go_default_test",
     srcs = [
         "assignments_test.go",
+        "attester_test.go",
         "server_test.go",
         "status_test.go",
     ],
