+++ beacon-chain/rpc/validator/attester.go
@@ -1,0 +1,116 @@
+package validator
+
+import (
+	"context"
+
+	"github.com/gogo/protobuf/proto"
+	ptypes "github.com/gogo/protobuf/types"
+	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
+	"github.com/prysmaticlabs/prysm/beacon-chain/cache"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/state"
+	"go.opencensus.io/trace"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+// GetAttestationData requests that the beacon node produce an attestation data object,
+// which the validator acting as an attester will then sign.
+func (vs *Server) GetAttestationData(ctx context.Context, req *ethpb.AttestationDataRequest) (*ethpb.AttestationData, error) {
+	ctx, span := trace.StartSpan(ctx, "AttesterServer.RequestAttestation")
+	defer span.End()
+	span.AddAttributes(
+		trace.Int64Attribute("slot", int64(req.Slot)),
+		trace.Int64Attribute("committeeIndex", int64(req.CommitteeIndex)),
+	)
+
+	if vs.SyncChecker.Syncing() {
+		return nil, status.Errorf(codes.Unavailable, "Syncing to latest head, not ready to respond")
+	}
+
+	res, err := vs.AttestationCache.Get(ctx, req)
+	if err != nil {
+		return nil, status.Errorf(codes.Internal, "Could not retrieve data from attestation cache: %v", err)
+	}
+	if res != nil {
+		return res, nil
+	}
+
+	if err := vs.AttestationCache.MarkInProgress(req); err != nil {
+		if err == cache.ErrAlreadyInProgress {
+			res, err := vs.AttestationCache.Get(ctx, req)
+			if err != nil {
+				return nil, status.Errorf(codes.Internal, "Could not retrieve data from attestation cache: %v", err)
+			}
+			if res == nil {
+				return nil, status.Error(codes.DataLoss, "A request was in progress and resolved to nil")
+			}
+			return res, nil
+		}
+		return nil, status.Errorf(codes.Internal, "Could not mark attestation as in-progress: %v", err)
+	}
+	defer func() {
+		if err := vs.AttestationCache.MarkNotInProgress(req); err != nil {
+			log.WithError(err).Error("Failed to mark cache not in progress")
+		}
+	}()
+
+	headState, err := vs.HeadFetcher.HeadState(ctx)
+	if err != nil {
+		return nil, status.Errorf(codes.Internal, "Could not retrieve head state: %v", err)
+	}
+	headRoot := vs.HeadFetcher.HeadRoot()
+
+	headState, err = state.ProcessSlots(ctx, headState, req.Slot)
+	if err != nil {
+		return nil, status.Errorf(codes.Internal, "Could not process slots up to %d: %v", req.Slot, err)
+	}
+
+	targetEpoch := helpers.CurrentEpoch(headState)
+	epochStartSlot := helpers.StartSlot(targetEpoch)
+	targetRoot := make([]byte, 32)
+	if epochStartSlot == headState.Slot {
+		targetRoot = headRoot[:]
+	} else {
+		targetRoot, err = helpers.BlockRootAtSlot(headState, epochStartSlot)
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "Could not get target block for slot %d: %v", epochStartSlot, err)
+		}
+	}
+
+	res = &ethpb.AttestationData{
+		Slot:            req.Slot,
+		CommitteeIndex:  req.CommitteeIndex,
+		BeaconBlockRoot: headRoot[:],
+		Source:          headState.CurrentJustifiedCheckpoint,
+		Target: &ethpb.Checkpoint{
+			Epoch: targetEpoch,
+			Root:  targetRoot,
+		},
+	}
+
+	if err := vs.AttestationCache.Put(ctx, req, res); err != nil {
+		return nil, status.Errorf(codes.Internal, "Could not store attestation data in cache: %v", err)
+	}
+	return res, nil
+}
+
+// ProposeAttestation is a function called by an attester to vote
+// on a block via an attestation object as defined in the Ethereum Serenity specification.
+func (vs *Server) ProposeAttestation(ctx context.Context, att *ethpb.Attestation) (*ptypes.Empty, error) {
+	// Broadcast the new attestation to the network.
+	if err := vs.P2P.Broadcast(ctx, att); err != nil {
+		return nil, status.Errorf(codes.Internal, "Could not broadcast attestation: %v", err)
+	}
+
+	go func() {
+		ctx = trace.NewContext(context.Background(), trace.FromContext(ctx))
+		attCopy := proto.Clone(att).(*ethpb.Attestation)
+		if err := vs.OperationsHandler.HandleAttestation(ctx, attCopy); err != nil {
+			log.WithError(err).Error("Could not handle attestation in operations service")
+			return
+		}
+	}()
+
+	return &ptypes.Empty{}, nil
+}
